+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v0.2.0] - 2016-09-19
+
+Rename repositories to "API"
+
+### Changed
+- Rename sub APIs for workspaces, clients, projects and time entries from "...Repository" to "...API"
+
 ## [v0.1.3] - 2016-09-18
 
 Fix errors in example code